requirements.txt
django
djangorestframework
django-extensions
django_rq
django-environ
django-oauth-toolkit
django-cors-headers

psycopg2
shortuuid
netaddr
pynetbox
cerberus
proxmoxer
python-freeipa
towerlib
paramiko
scp